requirements.txt
asgiref==3.8.1
cffi==1.17.0
cryptography==43.0.0
Django==5.1
django-dotenv==1.4.2
django-environ==0.11.2
djangorestframework==3.15.2
psycopg2-binary==2.9.9
pycparser==2.22
python-dotenv==1.0.1
sqlparse==0.5.1
tzdata==2024.1
dj-database-url==1.0.0